+++ include/cowel/directive_behavior.hpp
@@ -114,6 +114,25 @@ protected:
     do_evaluate(const Invocation& call, Context& context) const = 0;
 };
 
+struct Int_Directive_Behavior : Directive_Behavior {
+    [[nodiscard]]
+    constexpr Int_Directive_Behavior() noexcept
+        : Directive_Behavior { Type::integer }
+    {
+    }
+
+    [[nodiscard]]
+    Result<Value, Processing_Status> evaluate(const Invocation& call, Context& context) const final;
+    [[nodiscard]]
+    Processing_Status
+    splice(Content_Policy& out, const Invocation& call, Context& context) const final;
+
+protected:
+    [[nodiscard]]
+    virtual Result<Integer, Processing_Status>
+    do_evaluate(const Invocation& call, Context& context) const = 0;
+};
+
 /// @brief The behavior of a directive that returns a value of type `str`
 /// which is guaranteed to fit into a `Short_String_Value`.
 ///