+++ src/main/cpp/directives/variables.cpp
@@ -1,3 +1,5 @@
+#include <cmath>
+#include <ranges>
 #include <string>
 #include <string_view>
 #include <utility>
@@ -5,6 +7,7 @@
 
 #include "cowel/util/assert.hpp"
 #include "cowel/util/char_sequence.hpp"
+#include "cowel/util/math.hpp"
 #include "cowel/util/result.hpp"
 
 #include "cowel/policy/content_policy.hpp"
@@ -21,39 +24,6 @@
 
 namespace cowel {
 namespace {
-
-// FIXME: the whole neutral element picking is technically dead code right now.
-//        Specifically, this happens because we always require nonempty packs,
-//        which is necessary because it would be unclear which type of zero cowel_add()
-//        should return otherwise.
-//        However, it is plausible that something like cowel_add_f32()
-//        may clarify that in the future.
-
-[[nodiscard]]
-int expression_type_neutral_element(Numeric_Expression_Kind e)
-{
-    switch (e) {
-    case Numeric_Expression_Kind::neg:
-    case Numeric_Expression_Kind::add:
-    case Numeric_Expression_Kind::sub: return 0;
-    case Numeric_Expression_Kind::mul:
-    case Numeric_Expression_Kind::div: return 1;
-    }
-    COWEL_ASSERT_UNREACHABLE(u8"Invalid expression type.");
-}
-
-[[maybe_unused]] [[nodiscard]]
-Value expression_type_neutral_element(Numeric_Expression_Kind e, const Type& type)
-{
-    const int result = expression_type_neutral_element(e);
-    switch (type.get_kind()) {
-    case Type_Kind::integer: return Value::integer(result);
-    case Type_Kind::f32: return Value::f32(Float32(result));
-    case Type_Kind::f64: return Value::f64(Float64(result));
-    default: break;
-    }
-    COWEL_ASSERT_UNREACHABLE(u8"Unexpected type.");
-}
 
 [[nodiscard]] [[maybe_unused]]
 std::pmr::string vec_to_string(const std::pmr::vector<char>& v)
@@ -125,15 +95,75 @@ bool compare(const Value& x, const Value& y)
 
 template <typename T>
 [[nodiscard]]
-T operate_binary(Numeric_Expression_Kind type, T x, T y)
+T operate_unary(Unary_Numeric_Expression_Kind type, T x)
+{
+    if constexpr (std::is_integral_v<T>) {
+        switch (type) {
+        case Unary_Numeric_Expression_Kind::pos: return +x;
+        case Unary_Numeric_Expression_Kind::neg: return -x;
+        case Unary_Numeric_Expression_Kind::abs: return std::abs(x);
+        default: COWEL_ASSERT_UNREACHABLE(u8"Invalid unary operation for integers.");
+        }
+    }
+    else {
+        switch (type) {
+        case Unary_Numeric_Expression_Kind::pos: return +x;
+        case Unary_Numeric_Expression_Kind::neg: return -x;
+        case Unary_Numeric_Expression_Kind::abs: return std::abs(x);
+        case Unary_Numeric_Expression_Kind::sqrt: return std::sqrt(x);
+        case Unary_Numeric_Expression_Kind::trunc: return std::trunc(x);
+        case Unary_Numeric_Expression_Kind::floor: return std::floor(x);
+        case Unary_Numeric_Expression_Kind::ceil: return std::ceil(x);
+        case Unary_Numeric_Expression_Kind::nearest: return cowel::roundeven(x);
+        case Unary_Numeric_Expression_Kind::nearest_away_zero: return std::round(x);
+        }
+    }
+    COWEL_ASSERT_UNREACHABLE(u8"Invalid expression type.");
+}
+
+template <typename T>
+[[nodiscard]]
+T operate_binary(N_Ary_Numeric_Expression_Kind type, T x, T y)
 {
     switch (type) {
-    case Numeric_Expression_Kind::neg:
-        COWEL_ASSERT_UNREACHABLE(u8"Negation is not a binary operation.");
-    case Numeric_Expression_Kind::add: return x + y;
-    case Numeric_Expression_Kind::sub: return x - y;
-    case Numeric_Expression_Kind::mul: return x * y;
-    case Numeric_Expression_Kind::div: return x / y;
+    case N_Ary_Numeric_Expression_Kind::add: return x + y;
+    case N_Ary_Numeric_Expression_Kind::sub: return x - y;
+    case N_Ary_Numeric_Expression_Kind::mul: return x * y;
+    case N_Ary_Numeric_Expression_Kind::div: {
+        COWEL_DEBUG_ASSERT(y != 0);
+        return x / y;
+    }
+    case N_Ary_Numeric_Expression_Kind::min: {
+        if constexpr (std::is_floating_point_v<T>) {
+            return cowel::fminimum(x, y);
+        }
+        else {
+            return std::min(x, y);
+        }
+    }
+    case N_Ary_Numeric_Expression_Kind::max: {
+        if constexpr (std::is_floating_point_v<T>) {
+            return cowel::fmaximum(x, y);
+        }
+        else {
+            return std::max(x, y);
+        }
+    }
+    }
+    COWEL_ASSERT_UNREACHABLE(u8"Invalid expression type.");
+}
+
+[[nodiscard]]
+Integer operate_binary(Integer_Division_Kind type, Integer x, Integer y)
+{
+    COWEL_DEBUG_ASSERT(y != 0);
+    switch (type) {
+    case Integer_Division_Kind::div_to_zero: return x / y;
+    case Integer_Division_Kind::rem_to_zero: return x % y;
+    case Integer_Division_Kind::div_to_pos_inf: return div_to_pos_inf(x, y);
+    case Integer_Division_Kind::rem_to_pos_inf: return rem_to_pos_inf(x, y);
+    case Integer_Division_Kind::div_to_neg_inf: return div_to_neg_inf(x, y);
+    case Integer_Division_Kind::rem_to_neg_inf: return rem_to_neg_inf(x, y);
     }
     COWEL_ASSERT_UNREACHABLE(u8"Invalid expression type.");
 }
@@ -324,11 +354,99 @@ Comparison_Expression_Behavior::do_evaluate(const Invocation& call, Context& con
 }
 
 Result<Value, Processing_Status>
-Numeric_Expression_Behavior::evaluate(const Invocation& call, Context& context) const
+Unary_Numeric_Expression_Behavior::evaluate(const Invocation& call, Context& context) const
 {
     static const Type numeric_type
         = Type::canonical_union_of({ Type::integer, Type::f32, Type::f64 });
 
+    Group_Pack_Value_Matcher group_matcher { context.get_transient_memory() };
+    Call_Matcher call_matcher { group_matcher };
+
+    const auto match_status = call_matcher.match_call(call, context, make_fail_callback());
+    if (match_status != Processing_Status::ok) {
+        return match_status;
+    }
+
+    if (group_matcher.get_values().size() != 1) {
+        context.try_error(
+            diagnostic::type_mismatch, call.get_arguments_source_span(),
+            u8"Unary operation requires exactly one argument"sv
+        );
+        return Processing_Status::error;
+    }
+
+    const auto& first_value = group_matcher.get_values().front().value;
+    const auto& first_type = first_value.get_type();
+
+    switch (first_type.get_kind()) {
+    case Type_Kind::integer: {
+        return Value::integer(operate_unary(m_expression_kind, first_value.as_integer()));
+    }
+    case Type_Kind::f32: {
+        return Value::f32(operate_unary(m_expression_kind, first_value.as_f32()));
+    }
+    case Type_Kind::f64: {
+        return Value::f64(operate_unary(m_expression_kind, first_value.as_f64()));
+    }
+    default: break;
+    }
+    COWEL_ASSERT_UNREACHABLE(u8"Type of value should have already been checked.");
+}
+
+Result<Integer, Processing_Status>
+Integer_Division_Expression_Behavior::do_evaluate(const Invocation& call, Context& context) const
+{
+    Group_Pack_Value_Matcher group_matcher { context.get_transient_memory() };
+    Call_Matcher call_matcher { group_matcher };
+
+    const auto match_status = call_matcher.match_call(call, context, make_fail_callback());
+    if (match_status != Processing_Status::ok) {
+        return match_status;
+    }
+
+    bool type_check_ok = true;
+    if (group_matcher.get_values().size() != 2) {
+        context.try_error(
+            diagnostic::type_mismatch, call.get_arguments_source_span(),
+            u8"Binary operation requires two arguments."sv
+        );
+        type_check_ok = false;
+    }
+
+    for (const auto& [value, location] : group_matcher.get_values()) {
+        if (value.get_type() != Type::integer) {
+            context.try_error(
+                diagnostic::type_mismatch, location,
+                joined_char_sequence({
+                    u8"Expected a value of type "sv,
+                    Type::integer.get_display_name(),
+                    u8", but got "sv,
+                    value.get_type().get_display_name(),
+                    u8".",
+                })
+            );
+            type_check_ok = false;
+        }
+    }
+    if (!type_check_ok) {
+        return Processing_Status::error;
+    }
+
+    const auto& x_value = group_matcher.get_values()[0].value;
+    const auto& [y_value, y_location] = group_matcher.get_values()[1];
+
+    if (y_value.as_integer() == 0) {
+        context.try_error(diagnostic::arithmetic_div_by_zero, y_location, u8"Division by zero."sv);
+        return Processing_Status::error;
+    }
+    const Integer result
+        = operate_binary(m_expression_kind, x_value.as_integer(), y_value.as_integer());
+    return result;
+}
+
+Result<Value, Processing_Status>
+N_Ary_Numeric_Expression_Behavior::evaluate(const Invocation& call, Context& context) const
+{
     Group_Pack_Value_Matcher group_matcher { context.get_transient_memory() };
     Call_Matcher call_matcher { group_matcher };
 
@@ -348,42 +466,40 @@ Numeric_Expression_Behavior::evaluate(const Invocation& call, Context& context) 
     const auto& first_value = group_matcher.get_values().front().value;
     const auto& first_type = first_value.get_type();
 
-    if (m_expression_kind == Numeric_Expression_Kind::neg) {
-        if (group_matcher.get_values().size() != 1) {
-            context.try_error(
-                diagnostic::type_mismatch, call.get_arguments_source_span(),
-                u8"Negation is unary and requires exactly one argument"sv
-            );
-            return Processing_Status::error;
-        }
-        switch (first_type.get_kind()) {
-        case Type_Kind::integer: {
-            return Value::integer(-first_value.as_integer());
-        }
-        case Type_Kind::f32: {
-            return Value::f32(-first_value.as_f32());
-        }
-        case Type_Kind::f64: {
-            return Value::f64(-first_value.as_f64());
-        }
-        default: break;
-        }
-        COWEL_ASSERT_UNREACHABLE(u8"Type of value should have already been checked.");
-    }
-
-    // Type checks and check for integer division by zero.
+    bool type_check_ok = true;
     for (const auto& [value, location] : group_matcher.get_values()) {
-        if (!value.get_type().analytically_convertible_to(numeric_type)) {
-            context.try_error(
-                diagnostic::type_mismatch, location,
-                joined_char_sequence({
-                    u8"Expected a value of type "sv,
-                    numeric_type.get_display_name(),
-                    u8", but got "sv,
-                    value.get_type().get_display_name(),
-                    u8".",
-                })
-            );
+        if (m_expression_kind == N_Ary_Numeric_Expression_Kind::div) {
+            static const Type floating_type = Type::canonical_union_of({ Type::f32, Type::f64 });
+
+            if (!value.get_type().analytically_convertible_to(floating_type)) {
+                context.try_error(
+                    diagnostic::type_mismatch, location,
+                    joined_char_sequence({
+                        u8"Expected a value of floating-point type, but got "sv,
+                        value.get_type().get_display_name(),
+                        u8".",
+                    })
+                );
+                type_check_ok = false;
+            }
+        }
+        else {
+            static const Type numeric_type
+                = Type::canonical_union_of({ Type::integer, Type::f32, Type::f64 });
+
+            if (!value.get_type().analytically_convertible_to(numeric_type)) {
+                context.try_error(
+                    diagnostic::type_mismatch, location,
+                    joined_char_sequence({
+                        u8"Expected a value of type "sv,
+                        numeric_type.get_display_name(),
+                        u8", but got "sv,
+                        value.get_type().get_display_name(),
+                        u8".",
+                    })
+                );
+                type_check_ok = false;
+            }
         }
         if (value.get_type() != first_type) {
             context.try_error(
@@ -394,56 +510,38 @@ Numeric_Expression_Behavior::evaluate(const Invocation& call, Context& context) 
                     u8".",
                 })
             );
+            type_check_ok = false;
         }
     }
+    if (!type_check_ok) {
+        return Processing_Status::error;
+    }
 
-    bool first = true;
+    COWEL_ASSERT(!group_matcher.get_values().empty());
+
+    const auto values_without_first = group_matcher.get_values() | std::views::drop(1);
+
     switch (first_type.get_kind()) {
     case Type_Kind::integer: {
-        auto result = Integer(expression_type_neutral_element(m_expression_kind));
-        for (const auto& [value, location] : group_matcher.get_values()) {
-            if (first) {
-                first = false;
-                result = value.as_integer();
-            }
-            else if (m_expression_kind == Numeric_Expression_Kind::div && value.as_integer() == 0) {
-                context.try_error(
-                    diagnostic::arithmetic_div_by_zero, location, u8"Division by zero."sv
-                );
-                return Processing_Status::error;
-            }
-            else {
-
-                result = operate_binary(m_expression_kind, result, value.as_integer());
-            }
+        Integer result = first_value.as_integer();
+        for (const auto& [value, location] : values_without_first) {
+            result = operate_binary(m_expression_kind, result, value.as_integer());
         }
         return Value::integer(result);
     }
 
     case Type_Kind::f32: {
-        auto result = Float32(expression_type_neutral_element(m_expression_kind));
-        for (const auto& [value, location] : group_matcher.get_values()) {
-            if (first) {
-                first = false;
-                result = value.as_f32();
-            }
-            else {
-                result = operate_binary(m_expression_kind, result, value.as_f32());
-            }
+        Float32 result = first_value.as_f32();
+        for (const auto& [value, location] : values_without_first) {
+            result = operate_binary(m_expression_kind, result, value.as_f32());
         }
         return Value::f32(result);
     }
 
     case Type_Kind::f64: {
-        auto result = Float64(expression_type_neutral_element(m_expression_kind));
-        for (const auto& [value, location] : group_matcher.get_values()) {
-            if (first) {
-                first = false;
-                result = value.as_f64();
-            }
-            else {
-                result = operate_binary(m_expression_kind, result, value.as_f64());
-            }
+        Float64 result = first_value.as_f64();
+        for (const auto& [value, location] : values_without_first) {
+            result = operate_binary(m_expression_kind, result, value.as_f64());
         }
         return Value::f64(result);
     }